hot-fix: hard-code data bitrate to 2'000'000
As it turns out: The QT Serial library for the Peak devices doesn't
support CAN-FD data bit rate to be below 2'000'000. The values are
hard-coded in the QT library. This is due to PCAN having a configuration
string with TSEG1, TSEG2 and prescaler which isn't dynamically derived
by QT, but instead has only a limited set of available hard-coded
options encoded as strings.

```
static QByteArray dataBitrateString(int dataBitrate)
{
    switch (dataBitrate) {
    case 2000000:
        return ", data_brp=4, data_tseg1=7, data_tseg2=2, data_sjw=1";
    …
```

Unlucky for us, our CAN-FD bus uses a data bitrate of 1MHz (i know). So
we will be choosing the next possible option so that the code path
configures CANFD on the PCAN. To still be able to speak with a 1MHz
CANFD bus, we will be editing the QT DLLs, so that it in actuality
configures the PCAN to speak CAN-FD 1MHz, using the following string
which will result in a 80% sample point on 1MHz.

```
, data_brp=8, data_tseg1=7, data_tseg2=2, data_sjw=1
```

Signed-off-by: Elisabeth Friedrich <auxsys@flowerpot.me>